requirements.txt
# utils
tqdm==4.64.1
dvc[gdrive]==2.10.2
jupyter==1.0.0
jupyterlab==3.5.0
joblib==1.2.0
neptune-client==0.16.13
typer==0.7.0
huggingface_hub==0.11.0
jsonnet

# data
pandas==1.3.5
numpy==1.21.6

# ml and stuff
scikit-learn==1.0.2
spacy==3.4.3
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113; sys_platform == "linux"
torch==1.12.1; sys_platform == "darwin"
pytorch-lightning==1.8.3.post1
fasttext==0.9.2
scikit-optimize==0.9.0
transformers==4.24.0
datasets==2.7.1
evaluate==0.3.0
torchinfo==1.7.1
accelerate==0.14.0
tensorboard==2.10.1
tokenizers
sentencepiece

# data processing
emoji==2.2.0
wordsegment==1.3.1

# eda
matplotlib==3.5.3
seaborn==0.12.1
